v0.2.0 Changed namespace name, breaking change
Namespace is now ca.grammati.daisy (I'm going to try to be consistent in
my open-source projects, using ca.grammati).

Added `system-map` function, which will try to deduce component
dependencies.

Added functions to start and stop systems with error handling.

Funcions to handle the commmon case of starting and stopping a system
and storing the result in a var.