v0.2.1 Don't assoc nil into a component
Component library gets cranky if it sees a nil.

Also be a bit more selective about extending daisy's Lifecycle, because
I was getting some stack overlows with daisy/start calling
component/start and v/v forever. I'm not super happy with the way this
works, and I need to think harder about how to make this robust, but for
now this is good enough.